+++ src/card/cards/vintage_cube.rs
@@ -188,3 +188,4 @@ pub const ABHORRENT_OCULUS: CardDefinitionId = CardDefinitionId(2270);
 pub const UNDERWORLD_BREACH: CardDefinitionId = CardDefinitionId(2271);
 pub const CRYPTIC_COMMAND: CardDefinitionId = CardDefinitionId(2272);
 pub const CATHAR_COMMANDO: CardDefinitionId = CardDefinitionId(2273);
+pub const ELITE_SPELLBINDER: CardDefinitionId = CardDefinitionId(2274);